Move Module system to module package
