+++ module/meta/ModuleBase.java
@@ -1,10 +1,11 @@
-package vague.module;
+package module.meta;
 
 import java.awt.Graphics;
 import java.awt.event.KeyEvent;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseWheelEvent;
 import java.awt.image.BufferedImage;
+import module.Module;
 import vague.util.Cursor;
 import vague.util.Vector;
