+++ vague/Window.java
@@ -26,11 +26,11 @@ import javax.swing.Timer;
 import vague.geom.Rectangle;
 import vague.input.Control;
 import vague.input.Controls;
-import vague.module.Exchange;
-import vague.module.GraphicsCallback;
-import vague.module.Module;
+import module.Exchange;
+import module.GraphicsCallback;
+import module.Module;
 import vague.module.TestModule;
-import vague.module.container.Container;
+import module.container.Container;
 import vague.util.MouseTracker;
 import vague.util.Vector;
 import vague.workspace.Workspace;
